[GTK][WPE] Merge CompositingCoordinator into LayerTreeHost https://bugs.webkit.org/show_bug.cgi?id=280965
Reviewed by Alejandro G. Castro.

* Source/WebKit/PlatformPlayStation.cmake:
* Source/WebKit/PlatformWin.cmake:
* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/CompositingCoordinator.cpp: Removed.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/CompositingCoordinator.h: Removed.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp:
(WebKit::LayerTreeHost::LayerTreeHost):
(WebKit::LayerTreeHost::~LayerTreeHost):
(WebKit::LayerTreeHost::flushLayers):
(WebKit::LayerTreeHost::layerFlushTimerFired):
(WebKit::LayerTreeHost::setRootCompositingLayer):
(WebKit::LayerTreeHost::setViewOverlayRootLayer):
(WebKit::LayerTreeHost::forceRepaint):
(WebKit::LayerTreeHost::sizeDidChange):
(WebKit::LayerTreeHost::graphicsLayerFactory):
(WebKit::LayerTreeHost::didChangeViewport):
(WebKit::LayerTreeHost::detachLayer):
(WebKit::LayerTreeHost::attachLayer):
(WebKit::LayerTreeHost::paintingEngine):
(WebKit::LayerTreeHost::imageBackingStore):
(WebKit::LayerTreeHost::createGraphicsLayer):
(WebKit::LayerTreeHost::requestDisplayRefreshMonitorUpdate):
(WebKit::LayerTreeHost::requestUpdate):
(WebKit::LayerTreeHost::renderNextFrame):
(WebKit::LayerTreeHost::didFlushRootLayer): Deleted.
(WebKit::LayerTreeHost::updateScene): Deleted.
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.h:
(WebKit::LayerTreeHost::LayerTreeHost): Deleted.
(WebKit::LayerTreeHost::~LayerTreeHost): Deleted.
(WebKit::LayerTreeHost::setLayerFlushSchedulingEnabled): Deleted.
(WebKit::LayerTreeHost::scheduleLayerFlush): Deleted.
(WebKit::LayerTreeHost::cancelPendingLayerFlush): Deleted.
(WebKit::LayerTreeHost::setRootCompositingLayer): Deleted.
(WebKit::LayerTreeHost::setViewOverlayRootLayer): Deleted.
(WebKit::LayerTreeHost::scrollNonCompositedContents): Deleted.
(WebKit::LayerTreeHost::forceRepaint): Deleted.
(WebKit::LayerTreeHost::forceRepaintAsync): Deleted.
(WebKit::LayerTreeHost::sizeDidChange): Deleted.
(WebKit::LayerTreeHost::pauseRendering): Deleted.
(WebKit::LayerTreeHost::resumeRendering): Deleted.
(WebKit::LayerTreeHost::graphicsLayerFactory): Deleted.
(WebKit::LayerTreeHost::contentsSizeChanged): Deleted.
(WebKit::LayerTreeHost::didChangeViewportAttributes): Deleted.
(WebKit::LayerTreeHost::deviceOrPageScaleFactorChanged): Deleted.
(WebKit::LayerTreeHost::adjustTransientZoom): Deleted.
(WebKit::LayerTreeHost::commitTransientZoom): Deleted.
(WebKit::LayerTreeHost::createDisplayRefreshMonitor): Deleted.

Canonical link: https://commits.webkit.org/284817@main
