+++ Source/WebKit/PlatformPlayStation.cmake
@@ -93,7 +93,6 @@ list(APPEND WebKit_SOURCES
 
     WebProcess/WebPage/AcceleratedSurface.cpp
 
-    WebProcess/WebPage/CoordinatedGraphics/CompositingCoordinator.cpp
     WebProcess/WebPage/CoordinatedGraphics/DrawingAreaCoordinatedGraphics.cpp
 
     WebProcess/WebPage/libwpe/AcceleratedSurfaceLibWPE.cpp
