+++ Source/WebKit/SourcesWPE.txt
@@ -288,7 +288,6 @@ WebProcess/WebCoreSupport/wpe/WebEditorClientWPE.cpp
 WebProcess/WebPage/AcceleratedSurface.cpp
 
 WebProcess/WebPage/CoordinatedGraphics/CoordinatedGraphicsScene.cpp
-WebProcess/WebPage/CoordinatedGraphics/CompositingCoordinator.cpp
 WebProcess/WebPage/CoordinatedGraphics/CompositingRunLoop.cpp
 WebProcess/WebPage/CoordinatedGraphics/DrawingAreaCoordinatedGraphics.cpp
 WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp
