+++ Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.h
@@ -26,13 +26,20 @@
 
 #pragma once
 
+#if USE(COORDINATED_GRAPHICS)
 #include "AcceleratedSurface.h"
 #include "CallbackID.h"
-#include "CompositingCoordinator.h"
 #include "LayerTreeContext.h"
 #include "SimpleViewportController.h"
 #include "ThreadedCompositor.h"
+#include <WebCore/CoordinatedGraphicsLayer.h>
+#include <WebCore/CoordinatedImageBackingStore.h>
 #include <WebCore/FloatPoint.h>
+#include <WebCore/GraphicsLayerClient.h>
+#include <WebCore/GraphicsLayerFactory.h>
+#include <WebCore/NicosiaPlatformLayer.h>
+#include <WebCore/NicosiaScene.h>
+#include <WebCore/NicosiaSceneIntegration.h>
 #include <WebCore/PlatformScreen.h>
 #include <wtf/Forward.h>
 #include <wtf/OptionSet.h>
@@ -43,11 +50,10 @@
 #include "ThreadedDisplayRefreshMonitor.h"
 #endif
 
-#if USE(GRAPHICS_LAYER_TEXTURE_MAPPER)
-
-#include "LayerTreeHostTextureMapper.h"
-
-#else // USE(GRAPHICS_LAYER_TEXTURE_MAPPER)
+namespace Nicosia {
+class PaintingEngine;
+class SceneIntegration;
+}
 
 namespace WebCore {
 class CoordinatedGraphicsLayer;
@@ -56,6 +62,8 @@ class IntRect;
 class IntSize;
 class GraphicsLayer;
 class GraphicsLayerFactory;
+class NativeImage;
+class SkiaThreadedPaintingPool;
 struct ViewportAttributes;
 }
 
@@ -63,12 +71,9 @@ namespace WebKit {
 
 class WebPage;
 
-class LayerTreeHost
-#if USE(COORDINATED_GRAPHICS)
-    final : public CompositingCoordinator::Client, public AcceleratedSurface::Client, public ThreadedCompositor::Client
+class LayerTreeHost final : public WebCore::GraphicsLayerClient, public WebCore::CoordinatedGraphicsLayerClient, public WebCore::GraphicsLayerFactory, public AcceleratedSurface::Client, public ThreadedCompositor::Client, public Nicosia::SceneIntegration::Client
 #if !HAVE(DISPLAY_LINK)
     , public ThreadedDisplayRefreshMonitor::Client
-#endif
 #endif
 {
     WTF_MAKE_TZONE_ALLOCATED(LayerTreeHost);
@@ -118,18 +123,30 @@ public:
     void preferredBufferFormatsDidChange();
 #endif
 private:
-#if USE(COORDINATED_GRAPHICS)
     void layerFlushTimerFired();
+    void flushLayers();
+    void commitSceneState(const RefPtr<Nicosia::Scene>&);
     void didChangeViewport();
 #if HAVE(DISPLAY_LINK)
     void didRenderFrameTimerFired();
 #endif
     void renderNextFrame(bool);
 
-    // CompositingCoordinator::Client
-    void didFlushRootLayer(const WebCore::FloatRect& visibleContentRect) override;
-    void commitSceneState(const RefPtr<Nicosia::Scene>&) override;
-    void updateScene() override;
+    // CoordinatedGraphicsLayerClient
+    bool isFlushingLayerChanges() const override { return m_isFlushingLayerChanges; }
+    WebCore::FloatRect visibleContentsRect() const override { return m_visibleContentsRect; }
+    void detachLayer(WebCore::CoordinatedGraphicsLayer*) override;
+    void attachLayer(WebCore::CoordinatedGraphicsLayer*) override;
+#if USE(CAIRO)
+    Nicosia::PaintingEngine& paintingEngine() override;
+#elif USE(SKIA)
+    WebCore::BitmapTexturePool* skiaAcceleratedBitmapTexturePool() const override { return m_skiaAcceleratedBitmapTexturePool.get(); }
+    WebCore::SkiaThreadedPaintingPool* skiaThreadedPaintingPool() const override { return m_skiaThreadedPaintingPool.get(); }
+#endif
+    Ref<WebCore::CoordinatedImageBackingStore> imageBackingStore(Ref<WebCore::NativeImage>&&) override;
+
+    // GraphicsLayerFactory
+    Ref<WebCore::GraphicsLayer> createGraphicsLayer(WebCore::GraphicsLayer::Type, WebCore::GraphicsLayerClient&) override;
 
     // AcceleratedSurface::Client
     void frameComplete() override;
@@ -145,6 +162,9 @@ private:
     void didRenderFrame(uint32_t, const WebCore::Damage&) override;
     void displayDidRefresh(WebCore::PlatformDisplayID) override;
 
+    // Nicosia::SceneIntegration::Client
+    void requestUpdate() override;
+
 #if !HAVE(DISPLAY_LINK)
     // ThreadedDisplayRefreshMonitor::Client
     void requestDisplayRefreshMonitorUpdate() override;
@@ -157,22 +177,28 @@ private:
     void applyTransientZoomToLayers(double, WebCore::FloatPoint);
 #endif
 
-#endif // USE(COORDINATED_GRAPHICS)
-
     WebPage& m_webPage;
     LayerTreeContext m_layerTreeContext;
-#if USE(COORDINATED_GRAPHICS)
+    RefPtr<WebCore::GraphicsLayer> m_rootLayer;
+    WebCore::GraphicsLayer* m_rootCompositingLayer { nullptr };
+    WebCore::GraphicsLayer* m_overlayCompositingLayer { nullptr };
+    HashMap<Nicosia::PlatformLayer::LayerID, WebCore::CoordinatedGraphicsLayer*> m_registeredLayers;
+    bool m_didInitializeRootCompositingLayer { false };
     bool m_layerFlushSchedulingEnabled { true };
+    bool m_isFlushingLayerChanges { false };
+    bool m_isPurgingBackingStores { false };
     bool m_isSuspended { false };
     bool m_isWaitingForRenderer { false };
     bool m_scheduledWhileWaitingForRenderer { false };
+    bool m_forceFrameSync { false };
     float m_lastPageScaleFactor { 1 };
     WebCore::IntPoint m_lastScrollPosition;
     bool m_scrolledSinceLastFrame { false };
-    WebCore::GraphicsLayer* m_viewOverlayRootLayer { nullptr };
+    double m_lastAnimationServiceTime { 0 };
     std::unique_ptr<AcceleratedSurface> m_surface;
     RefPtr<ThreadedCompositor> m_compositor;
     SimpleViewportController m_viewportController;
+    WebCore::FloatRect m_visibleContentsRect;
     struct {
         CompletionHandler<void()> callback;
         bool needsFreshFlush { false };
@@ -180,12 +206,21 @@ private:
     RunLoop::Timer m_layerFlushTimer;
 #if HAVE(DISPLAY_LINK)
     RunLoop::Timer m_didRenderFrameTimer;
-#endif
-    CompositingCoordinator m_coordinator;
-#endif // USE(COORDINATED_GRAPHICS)
-#if !HAVE(DISPLAY_LINK)
+#else
     WebCore::PlatformDisplayID m_displayID;
 #endif
+#if USE(CAIRO)
+    std::unique_ptr<Nicosia::PaintingEngine> m_paintingEngine;
+#elif USE(SKIA)
+    std::unique_ptr<WebCore::BitmapTexturePool> m_skiaAcceleratedBitmapTexturePool;
+    std::unique_ptr<WebCore::SkiaThreadedPaintingPool> m_skiaThreadedPaintingPool;
+#endif
+    HashMap<uint64_t, Ref<WebCore::CoordinatedImageBackingStore>> m_imageBackingStores;
+    struct {
+        RefPtr<Nicosia::Scene> scene;
+        RefPtr<Nicosia::SceneIntegration> sceneIntegration;
+        Nicosia::Scene::State state;
+    } m_nicosia;
 
 #if PLATFORM(GTK)
     bool m_transientZoom { false };
@@ -197,36 +232,6 @@ private:
     uint32_t m_compositionResponseID { 0 };
 };
 
-#if !USE(COORDINATED_GRAPHICS)
-#if HAVE(DISPLAY_LINK)
-inline LayerTreeHost::LayerTreeHost(WebPage& webPage) : m_webPage(webPage) { }
-#else
-inline LayerTreeHost::LayerTreeHost(WebPage& webPage, WebCore::PlatformDisplayID displayID) : m_webPage(webPage), m_displayID(displayID) { }
-#endif
-inline LayerTreeHost::~LayerTreeHost() { }
-inline void LayerTreeHost::setLayerFlushSchedulingEnabled(bool) { }
-inline void LayerTreeHost::scheduleLayerFlush() { }
-inline void LayerTreeHost::cancelPendingLayerFlush() { }
-inline void LayerTreeHost::setRootCompositingLayer(WebCore::GraphicsLayer*) { }
-inline void LayerTreeHost::setViewOverlayRootLayer(WebCore::GraphicsLayer*) { }
-inline void LayerTreeHost::scrollNonCompositedContents(const WebCore::IntRect&) { }
-inline void LayerTreeHost::forceRepaint() { }
-inline void LayerTreeHost::forceRepaintAsync(CompletionHandler<void()>&&) { }
-inline void LayerTreeHost::sizeDidChange(const WebCore::IntSize&) { }
-inline void LayerTreeHost::pauseRendering() { }
-inline void LayerTreeHost::resumeRendering() { }
-inline WebCore::GraphicsLayerFactory* LayerTreeHost::graphicsLayerFactory() { return nullptr; }
-inline void LayerTreeHost::contentsSizeChanged(const WebCore::IntSize&) { }
-inline void LayerTreeHost::didChangeViewportAttributes(WebCore::ViewportAttributes&&) { }
-inline void LayerTreeHost::deviceOrPageScaleFactorChanged() { }
-#if PLATFORM(GTK)
-inline void LayerTreeHost::adjustTransientZoom(double, WebCore::FloatPoint) { }
-inline void LayerTreeHost::commitTransientZoom(double, WebCore::FloatPoint) { }
-#else
-inline RefPtr<WebCore::DisplayRefreshMonitor> LayerTreeHost::createDisplayRefreshMonitor(WebCore::PlatformDisplayID) { return nullptr; }
-#endif
-#endif
-
 } // namespace WebKit
 
-#endif // USE(GRAPHICS_LAYER_TEXTURE_MAPPER)
+#endif // USE(COORDINATED_GRAPHICS)
